Update qbeast tests

+++ src/test/scala/io/qbeast/QbeastIntegrationTestSpec.scala
@@ -21,8 +21,6 @@ import io.qbeast.context.QbeastContextImpl
 import io.qbeast.core.model.IndexManager
 import io.qbeast.core.model.QTableID
 import io.qbeast.core.model.QbeastSnapshot
-import io.qbeast.spark.delta.DeltaMetadataManager
-import io.qbeast.spark.delta.DeltaRollupDataWriter
 import io.qbeast.spark.index.SparkColumnsToIndexSelector
 import io.qbeast.spark.index.SparkOTreeManager
 import io.qbeast.spark.index.SparkRevisionFactory
@@ -35,7 +33,9 @@ import org.apache.spark.SparkConf
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 
+import java.io.File
 import java.nio.file.Files
+import scala.reflect.io.Directory
 
 /**
  * This class contains all function that you should use to test qbeast over spark. You can use it
@@ -86,6 +86,7 @@ trait QbeastIntegrationTestSpec extends AnyFlatSpec with Matchers with DatasetCo
    * @param testCode
    *   the code to run within the spark session
    * @tparam T
+   *   the type of the output
    * @return
    */
   def withExtendedSpark[T](sparkConf: SparkConf = new SparkConf())(
@@ -113,6 +114,7 @@ trait QbeastIntegrationTestSpec extends AnyFlatSpec with Matchers with DatasetCo
    * @param testCode
    *   the code to test within the spark session
    * @tparam T
+   *   the type of the output
    * @return
    */
   def withSpark[T](testCode: SparkSession => T): T = {
@@ -123,7 +125,9 @@ trait QbeastIntegrationTestSpec extends AnyFlatSpec with Matchers with DatasetCo
    * Runs code with a Temporary Directory. After execution, the content of the directory is
    * deleted.
    * @param testCode
+   *   the test code
    * @tparam T
+   *   the test result type
    * @return
    */
   def withTmpDir[T](testCode: String => T): T = {
@@ -131,8 +135,14 @@ trait QbeastIntegrationTestSpec extends AnyFlatSpec with Matchers with DatasetCo
     try {
       testCode(directory.toString)
     } finally {
-      import scala.reflect.io.Directory
-      val d = new Directory(directory.toFile)
+      removeDirectory(directory.toString)
+    }
+  }
+
+  def removeDirectory(directoryPath: String): Unit = {
+    val directory = new File(directoryPath)
+    if (directory.exists() && directory.isDirectory) {
+      val d = new Directory(directory)
       d.deleteRecursively()
     }
   }
@@ -153,8 +163,8 @@ trait QbeastIntegrationTestSpec extends AnyFlatSpec with Matchers with DatasetCo
     withSpark { spark =>
       val indexedTableFactory = new IndexedTableFactoryImpl(
         SparkOTreeManager,
-        DeltaMetadataManager,
-        DeltaRollupDataWriter,
+        QbeastContext.metadataManager,
+        QbeastContext.dataWriter,
         SparkRevisionFactory,
         SparkColumnsToIndexSelector)
       val context = new QbeastContextImpl(spark.sparkContext.getConf, indexedTableFactory)
@@ -170,7 +180,9 @@ trait QbeastIntegrationTestSpec extends AnyFlatSpec with Matchers with DatasetCo
   /**
    * Runs code with QbeastContext and a Temporary Directory.
    * @param testCode
+   *   actual test code
    * @tparam T
+   *   output type
    * @return
    */
 
@@ -182,6 +194,7 @@ trait QbeastIntegrationTestSpec extends AnyFlatSpec with Matchers with DatasetCo
    * @param testCode
    *   the code to reproduce
    * @tparam T
+   *   the type of the output
    * @return
    */
   def withQbeastContextSparkAndTmpWarehouse[T](testCode: (SparkSession, String) => T): T =
@@ -193,7 +206,9 @@ trait QbeastIntegrationTestSpec extends AnyFlatSpec with Matchers with DatasetCo
   /**
    * Runs code with OTreeAlgorithm configuration
    * @param code
+   *   IndexManager
    * @tparam T
+   *   code type
    * @return
    */
   def withOTreeAlgorithm[T](code: IndexManager => T): T = {
